requirements.txt
# Core dependencies
hyperliquid-python-sdk>=0.9.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Web server for webhooks
fastapi>=0.100.0
uvicorn>=0.23.0

# HTTP clients
requests>=2.31.0
aiohttp>=3.8.0
httpx>=0.25.0

# Environment and config
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Visualization for backtesting
matplotlib>=3.7.0

# CLI
click>=8.0.0

# Logging
loguru>=0.7.0

# Async support
asyncio-throttle>=1.0.2
